refactor: moved timer views out of PDFView
